+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Updated dictionaries to [2025.06.30](https://github.com/grascii/dictionaries/tree/2025.06.30).
 - `grascii dephrase` is a first-class command
 - `dephrase` streams results in a more logical order
+- Added `AU`, `EU`, and `OE` to similarity graph with edge between `AU` and `EU`
 
 ### Removed